backend/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.8
Django==5.0.6
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.1
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==70.2.0
sqlparse==0.5.0
uritemplate==4.1.1
